move sortfield to config
