+++ src/components/Login/auth.js
@@ -4,8 +4,7 @@ function Auth(props) {
   const curURL = window.location.href;
   const url = new URL(curURL);
   let searchParams = new URLSearchParams(url.hash.split('?')[1]);
-  console.log(searchParams)
-  console.log(searchParams.get('access_token'))
+
   return (
     <>
       {<Navigate to="/main" />}
